// src/app/model/master/time-table-details.model.ts
export class TimeTableDetails {
  timeTableId: number;
  day: string;
  subject1: string;
  subject2: string;
  subject3: string;
  subject4: string;
  subject5: string;
  subject6: string;
  subject7: string;
  subject8: string;

  constructor() {
    this.timeTableId = 0;
    this.day = '';
    this.subject1 = '';
    this.subject2 = '';
    this.subject3 = '';
    this.subject4 = '';
    this.subject5 = '';
    this.subject6 = '';
    this.subject7 = '';
    this.subject8 = '';
  }
}